backend/requirements.txt
alembic==1.14.1
fastapi==0.115.8
httpx==0.28.1
psycopg[binary]==3.2.5
pydantic-settings==2.8.1
PyJWT==2.10.1
pytest==8.3.4
python-dotenv==1.0.1
SQLAlchemy==2.0.38
uvicorn[standard]==0.34.0

